// src/integrations/pms/abs-hotel/types.ts
export interface ABSReservationInput {

  reservationId: string;

  otaReferenceNumber?: string;

  editId?: string;

  calendarReservationId?: string;

  calendarStatus?: string;

  checkIn?: string;

  checkOut?: string;

  guestName?: string;

  firstName?: string;

  lastName?: string;

  email?: string;

  phone?: string;

  roomNumber?: string;

  roomType?: string;

  ratePlan?: string;

  hotelId?: string;

  hotelName?: string;

  revision?: number;

}